+++ libs/sdk-ui-ext/src/internal/components/pluggableVisualizations/pivotTable/PluggablePivotTable.tsx
@@ -2,32 +2,21 @@
 import cloneDeep from "lodash/cloneDeep";
 import flatMap from "lodash/flatMap";
 import get from "lodash/get";
-import includes from "lodash/includes";
 import isNil from "lodash/isNil";
 import isEmpty from "lodash/isEmpty";
+import isEqual from "lodash/isEqual";
 import { IExecutionFactory, ISettings, SettingCatalog } from "@gooddata/sdk-backend-spi";
 import {
-    attributeLocalId,
-    bucketAttributes,
-    bucketsFind,
-    bucketsMeasures,
-    IAttributeSortItem,
-    IBucket,
+    bucketAttribute,
     IDimension,
     IInsightDefinition,
-    IMeasureSortItem,
+    insightBucket,
     insightBuckets,
     insightHasDataDefined,
     insightProperties,
-    isAttributeSort,
-    isMeasureLocator,
-    isMeasureSort,
+    insightSorts,
     ISortItem,
     newAttributeSort,
-    measureLocalId,
-    insightSorts,
-    insightBucket,
-    bucketAttribute,
 } from "@gooddata/sdk-model";
 
 import { BucketNames, VisualizationEnvironment, VisualizationTypes } from "@gooddata/sdk-ui";
@@ -46,7 +35,6 @@ import { ATTRIBUTE, DATE, METRIC } from "../../../constants/bucket";
 import { DASHBOARDS_ENVIRONMENT } from "../../../constants/properties";
 import { DEFAULT_PIVOT_TABLE_UICONFIG } from "../../../constants/uiConfig";
 import {
-    IAttributeFilter,
     IBucketFilter,
     IBucketItem,
     IBucketOfFun,
@@ -64,7 +52,6 @@ import {
     getAllItemsByType,
     getItemsFromBuckets,
     getTotalsFromBucket,
-    isAttributeFilter,
     removeDuplicateBucketItems,
     sanitizeFilters,
 } from "../../../utils/bucketHelper";
@@ -79,7 +66,15 @@ import { setPivotTableUiConfig } from "../../../utils/uiConfigHelpers/pivotTable
 import UnsupportedConfigurationPanel from "../../configurationPanels/UnsupportedConfigurationPanel";
 import { AbstractPluggableVisualization } from "../AbstractPluggableVisualization";
 import { PIVOT_TABLE_SUPPORTED_PROPERTIES } from "../../../constants/supportedProperties";
-import { adaptReferencePointWidthItemsToPivotTable } from "./widthItemsHelpers";
+import {
+    adaptReferencePointWidthItemsToPivotTable,
+    adaptMdObjectWidthItemsToPivotTable,
+} from "./widthItemsHelpers";
+import {
+    adaptMdObjectSortItemsToPivotTable,
+    adaptReferencePointSortItemsToPivotTable,
+    addDefaultSort,
+} from "./sortItemsHelpers";
 
 export const getColumnAttributes = (buckets: IBucketOfFun[]): IBucketItem[] => {
     return getItemsFromBuckets(
@@ -102,181 +97,6 @@ export const getRowAttributes = (buckets: IBucketOfFun[]): IBucketItem[] => {
         [ATTRIBUTE, DATE],
     );
 };
-
-// removes attribute sortItems with invalid identifiers
-// removes measure sortItems with invalid identifiers and invalid number of locators
-function adaptSortItemsToPivotTable(
-    originalSortItems: ISortItem[],
-    measureLocalIdentifiers: string[],
-    rowAttributeLocalIdentifiers: string[],
-    columnAttributeLocalIdentifiers: string[],
-): ISortItem[] {
-    return originalSortItems.reduce((sortItems: ISortItem[], sortItem: ISortItem) => {
-        if (isMeasureSort(sortItem)) {
-            // filter out invalid locators
-            const filteredSortItem: IMeasureSortItem = {
-                measureSortItem: {
-                    ...sortItem.measureSortItem,
-                    locators: sortItem.measureSortItem.locators.filter((locator) => {
-                        // filter out invalid measure locators
-                        if (isMeasureLocator(locator)) {
-                            return includes(
-                                measureLocalIdentifiers,
-                                locator.measureLocatorItem.measureIdentifier,
-                            );
-                        }
-                        // filter out invalid column attribute locators
-                        return includes(
-                            columnAttributeLocalIdentifiers,
-                            locator.attributeLocatorItem.attributeIdentifier,
-                        );
-                    }),
-                },
-            };
-
-            // keep sortItem if measureLocator is present and locators are correct length
-            if (
-                filteredSortItem.measureSortItem.locators.some((locator) => isMeasureLocator(locator)) &&
-                filteredSortItem.measureSortItem.locators.length ===
-                    columnAttributeLocalIdentifiers.length + 1
-            ) {
-                return [...sortItems, filteredSortItem];
-            }
-
-            // otherwise just carry over previous sortItems
-            return sortItems;
-        }
-        /**
-         * Keep only row attribute sorts, column sorts are not supported.
-         *
-         * This exists to overcome AD weirdness where AD will sometimes send insight with invalid sorts
-         * even if the pivot table should NOT be sorted by default by the first row attribute in ascending order since it has no row attributes.
-         * Code here fixes this symptom and ensures the default sort is in place only if relevant.
-         *
-         * Typical case is PivotTable with one measure and one row and then the user moves thee attribute from Row bucket to Column bucket.
-         * In that case AD sends insight with the irrelevant sort and then without it.
-         *
-         * Note: while this may seem small thing, it's actually a messy business. When rendering / switching to the pivot
-         * table the AD will call update/render multiple times. Sometimes with sort items, sometimes without sort items. This
-         * can seriously mess up the pivot table in return: the column resizing is susceptible to race conditions and timing
-         * issues. Because of the flurry of calls, the table may not render or load indefinitely.
-         */
-        if (includes(rowAttributeLocalIdentifiers, sortItem.attributeSortItem.attributeIdentifier)) {
-            return [...sortItems, sortItem];
-        }
-        return sortItems;
-    }, []);
-}
-
-export function adaptReferencePointSortItemsToPivotTable(
-    originalSortItems: ISortItem[],
-    measures: IBucketItem[],
-    rowAttributes: IBucketItem[],
-    columnAttributes: IBucketItem[],
-): ISortItem[] {
-    const measureLocalIdentifiers = measures.map((measure) => measure.localIdentifier);
-    const rowAttributeLocalIdentifiers = rowAttributes.map((rowAttribute) => rowAttribute.localIdentifier);
-    const columnAttributeLocalIdentifiers = columnAttributes.map(
-        (columnAttribute) => columnAttribute.localIdentifier,
-    );
-
-    return adaptSortItemsToPivotTable(
-        originalSortItems,
-        measureLocalIdentifiers,
-        rowAttributeLocalIdentifiers,
-        columnAttributeLocalIdentifiers,
-    );
-}
-
-function adaptMdObjectSortItemsToPivotTable(originalSortItems: ISortItem[], buckets: IBucket[]): ISortItem[] {
-    const measureLocalIdentifiers = bucketsMeasures(buckets).map(measureLocalId);
-
-    const rowBucket = bucketsFind(buckets, BucketNames.ATTRIBUTE);
-    const rowAttributeLocalIdentifiers = rowBucket ? bucketAttributes(rowBucket).map(attributeLocalId) : [];
-
-    const columnBucket = bucketsFind(buckets, BucketNames.COLUMNS);
-    const columnAttributeLocalIdentifiers = columnBucket
-        ? bucketAttributes(columnBucket).map(attributeLocalId)
-        : [];
-
-    return adaptSortItemsToPivotTable(
-        originalSortItems,
-        measureLocalIdentifiers,
-        rowAttributeLocalIdentifiers,
-        columnAttributeLocalIdentifiers,
-    );
-}
-
-const isAttributeSortItemVisible = (_sortItem: IAttributeSortItem, _filters: IBucketFilter[]): boolean =>
-    true;
-
-const isMeasureSortItemMatchedByFilter = (sortItem: IMeasureSortItem, filter: IAttributeFilter): boolean =>
-    filter.selectedElements
-        ? filter.selectedElements.some((selectedElement) =>
-              sortItem.measureSortItem.locators.some(
-                  (locator) =>
-                      !isMeasureLocator(locator) &&
-                      locator.attributeLocatorItem.element === selectedElement.uri,
-              ),
-          )
-        : false;
-
-const isMeasureSortItemVisible = (sortItem: IMeasureSortItem, filters: IBucketFilter[]): boolean =>
-    filters.reduce((isVisible, filter) => {
-        if (isAttributeFilter(filter)) {
-            const shouldBeMatched = !filter.isInverted;
-            return isVisible && shouldBeMatched === isMeasureSortItemMatchedByFilter(sortItem, filter);
-        }
-        return isVisible;
-    }, true);
-
-export const isSortItemVisible = (sortItem: ISortItem, filters: IBucketFilter[]): boolean =>
-    isAttributeSort(sortItem)
-        ? isAttributeSortItemVisible(sortItem, filters)
-        : isMeasureSortItemVisible(sortItem, filters);
-
-export function addDefaultSort(
-    sortItems: ISortItem[],
-    filters: IBucketFilter[],
-    rowAttributes: IBucketItem[],
-    previousRowAttributes?: IBucketItem[],
-): ISortItem[] {
-    // cannot construct default sort without a row
-    if (rowAttributes.length < 1) {
-        return sortItems;
-    }
-
-    // detect custom sort
-    const firstRow = rowAttributes[0];
-    const previousFirstRow = previousRowAttributes && previousRowAttributes[0];
-    const hasVisibleCustomSort = sortItems.some((sortItem) => {
-        if (!isSortItemVisible(sortItem, filters)) {
-            return false;
-        }
-        // non attribute sort is definitely custom
-        if (!isAttributeSort(sortItem)) {
-            return true;
-        }
-        // asc sort on first row is considered default
-        if (
-            sortItem.attributeSortItem.attributeIdentifier === firstRow.localIdentifier &&
-            sortItem.attributeSortItem.direction === "asc"
-        ) {
-            return false;
-        }
-        // asc sort on row that was first until now is considered default as well
-        if (
-            previousFirstRow &&
-            sortItem.attributeSortItem.attributeIdentifier === previousFirstRow.localIdentifier &&
-            sortItem.attributeSortItem.direction === "asc"
-        ) {
-            return false;
-        }
-        return true;
-    });
-
-    return hasVisibleCustomSort ? sortItems : [newAttributeSort(firstRow.localIdentifier, "asc")];
-}
 
 export class PluggablePivotTable extends AbstractPluggableVisualization {
     private environment: VisualizationEnvironment;
@@ -415,6 +235,21 @@ export class PluggablePivotTable extends AbstractPluggableVisualization {
         };
     };
 
+    protected updateInstanceProperties(
+        options: IVisProps,
+        insight: IInsightDefinition,
+        insightPropertiesMeta: any,
+    ) {
+        super.updateInstanceProperties(options, insight, insightPropertiesMeta);
+
+        /*
+         * This was ported from v7. For some reason (likely related KD interop?) the entire content of properties
+         * would be picked up and used.
+         */
+        this.visualizationProperties = insightProperties(insight);
+        this.adaptPropertiesToInsight(this.visualizationProperties, this.currentInsight);
+    }
+
     protected renderVisualization(
         options: IVisProps,
         insight: IInsightDefinition,
@@ -529,31 +364,48 @@ export class PluggablePivotTable extends AbstractPluggableVisualization {
         return generateDimensions(insight, VisualizationTypes.TABLE);
     }
 
-    private getMergedProperties(newProperties: any): IVisualizationProperties {
-        const properties = this.visualizationProperties ?? {};
+    private adaptPropertiesToInsight(
+        visualizationProperties: IVisualizationProperties,
+        insight: IInsightDefinition,
+    ) {
+        // This is sanitization of properties from KD vs current mdObject from AD
+        const columnWidths = getColumnWidthsFromProperties(visualizationProperties);
+        if (columnWidths) {
+            this.sanitizeColumnWidths(columnWidths, insight);
+        }
+    }
 
-        return {
-            ...properties,
-            ...newProperties,
-        };
+    private sanitizeColumnWidths(columnWidths: ColumnWidthItem[], insight: IInsightDefinition) {
+        const adaptedColumnWidths = adaptMdObjectWidthItemsToPivotTable(columnWidths, insight);
+
+        if (!isEqual(columnWidths, adaptedColumnWidths)) {
+            this.visualizationProperties.properties.controls.columnWidths = adaptedColumnWidths;
+            this.pushData({
+                properties: {
+                    controls: {
+                        columnWidths: adaptedColumnWidths,
+                    },
+                },
+            });
+        }
     }
 
     private onColumnResized(columnWidths: ColumnWidthItem[]) {
         this.pushData({
-            properties: this.getMergedProperties({
+            properties: {
                 controls: {
                     columnWidths,
                 },
-            }),
+            },
         });
     }
 
     private handlePushData(data: any) {
         if (data && data.properties && data.properties.sortItems) {
             this.pushData({
-                properties: this.getMergedProperties({
+                properties: {
                     sortItems: data.properties.sortItems,
-                }),
+                },
             });
         } else {
             this.pushData(data);
