+++ libs/sdk-ui-ext/src/internal/components/pluggableVisualizations/pivotTable/tests/sortMocks.ts
@@ -1,0 +1,107 @@
+// (C) 2020 GoodData Corporation
+
+// sortItems based on referencePointMocks.simpleStackedReferencePoint
+import { IAttributeSortItem, IMeasureSortItem } from "@gooddata/sdk-model";
+
+export const validMeasureSort: IMeasureSortItem = {
+    measureSortItem: {
+        direction: "asc",
+        locators: [
+            {
+                attributeLocatorItem: {
+                    attributeIdentifier: "a2",
+                    element: "/gdc/md/PROJECTID/obj/2210/elements?id=1234",
+                },
+            },
+            {
+                measureLocatorItem: {
+                    measureIdentifier: "m1",
+                },
+            },
+        ],
+    },
+};
+export const validAttributeSort: IAttributeSortItem = {
+    attributeSortItem: {
+        attributeIdentifier: "a1",
+        direction: "desc",
+    },
+};
+export const invalidAttributeSort: IAttributeSortItem = {
+    attributeSortItem: {
+        attributeIdentifier: "invalid",
+        direction: "desc",
+    },
+};
+export const invalidMeasureSortInvalidMeasure: IMeasureSortItem = {
+    measureSortItem: {
+        direction: "asc",
+        locators: [
+            {
+                attributeLocatorItem: {
+                    attributeIdentifier: "a2",
+                    element: "/gdc/md/PROJECTID/obj/2210/elements?id=1234",
+                },
+            },
+            {
+                measureLocatorItem: {
+                    measureIdentifier: "invalid", // this measure is not in buckets
+                },
+            },
+        ],
+    },
+};
+export const invalidMeasureSortInvalidAttribute: IMeasureSortItem = {
+    measureSortItem: {
+        direction: "asc",
+        locators: [
+            {
+                attributeLocatorItem: {
+                    attributeIdentifier: "a1", // this identifier doesn't exist on second dimension
+                    element: "/gdc/md/PROJECTID/obj/2210/elements?id=1234",
+                },
+            },
+            {
+                measureLocatorItem: {
+                    measureIdentifier: "m1",
+                },
+            },
+        ],
+    },
+};
+export const invalidMeasureSortTooManyLocators: IMeasureSortItem = {
+    measureSortItem: {
+        direction: "asc",
+        locators: [
+            {
+                attributeLocatorItem: {
+                    attributeIdentifier: "a2",
+                    element: "/gdc/md/PROJECTID/obj/2210/elements?id=1234",
+                },
+            },
+            {
+                attributeLocatorItem: {
+                    attributeIdentifier: "a2",
+                    element: "/gdc/md/PROJECTID/obj/2210/elements?id=1234",
+                },
+            },
+            {
+                measureLocatorItem: {
+                    measureIdentifier: "m1",
+                },
+            },
+        ],
+    },
+};
+export const invalidMeasureSortLocatorsTooShort: IMeasureSortItem = {
+    measureSortItem: {
+        direction: "asc",
+        locators: [
+            {
+                measureLocatorItem: {
+                    measureIdentifier: "m1",
+                },
+            },
+        ],
+    },
+};
